requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==2.10.5
apache-airflow-providers-common-compat==1.7.0
apache-airflow-providers-common-io==1.6.0
apache-airflow-providers-common-sql==1.27.1
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.13.0
apache-airflow-providers-http==5.3.0
apache-airflow-providers-imap==3.9.0
apache-airflow-providers-smtp==2.1.0
apache-airflow-providers-sqlite==4.1.0
apispec==6.8.2
argcomplete==3.6.2
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
boto3==1.38.24
botocore==1.38.24
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clickclick==20.10.2
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==45.0.3
cycler==0.12.1
databricks-sdk==0.55.0
Deprecated==1.2.18
dill==0.4.0
dnspython==2.7.0
docker==7.1.0
email_validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
fire==0.7.0
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
google-re2==1.1.20240702
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.2
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
icecream==2.1.4
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy-object-proxy==1.11.0
limits==5.2.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.28.2
matplotlib==3.7.5
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
mlflow==2.22.0
mlflow-skinny==2.22.0
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
mysqlclient==2.2.7
networkx==3.4.2
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
ordered-set==4.1.0
packaging==24.2
pandas==2.0.3
pathspec==0.12.1
pendulum==3.1.0
pillow==11.2.1
pluggy==1.6.0
prison==0.2.1
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==7.4.4
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.1
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
scikit-learn==1.3.2
scipy==1.15.3
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
tomli==2.2.1
torch==2.4.1
tqdm==4.66.5
triton==3.0.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.4.0
uvicorn==0.32.1
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
WTForms==3.2.1
yarl==1.20.0
zipp==3.22.0
